test6

+++ src/main/java/com/weavusys/hrd/entity/Amount.java
@@ -1,8 +1,6 @@
 package com.weavusys.hrd.entity;
 
 import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
 import lombok.Data;
 import lombok.NoArgsConstructor;
@@ -13,6 +11,6 @@ import lombok.NoArgsConstructor;
 public class Amount {
 
     @Id
-    private Integer rank;  //직급 고유값
+    private Long rank;  //직급 고유값
     private Integer monthlyAmount;
 }
